Move root module to parent
