+++ backend/monaxia/src/worker.rs
@@ -1,5 +1,4 @@
 mod ap;
-mod root;
 mod user;
 
 use crate::{misc::create_http_client, repository_impl::construct_container_db};
@@ -9,15 +8,16 @@ use std::sync::Arc;
 use anyhow::Result;
 use lapin::{Connection as LapinConnection, ConnectionProperties};
 use monaxia_data::config::Config;
-use monaxia_job::job::MxJob;
+use monaxia_job::job::{Job, MxJob};
 use monaxia_queue::{
     job::{Consumer, Producer},
     queue::amqp::{create_amqp_consumer, create_amqp_producer},
 };
 use monaxia_repository::Container;
 use reqwest::Client;
+use serde_variant::to_variant_name;
 use tokio::spawn;
-use tracing::info;
+use tracing::{error, info, instrument};
 
 pub struct WorkerState {
     name: String,
@@ -32,13 +32,34 @@ struct JobState {
     pub http_client: Client,
 }
 
-pub async fn spawn_workers(config: Arc<Config>, consumers: Vec<Consumer<MxJob>>) -> Result<()> {
+pub async fn start_workers(config: Arc<Config>) -> Result<Producer<MxJob>> {
+    let conn = LapinConnection::connect(&config.queue.url, ConnectionProperties::default()).await?;
+    let producer = create_amqp_producer(&conn, "producer").await?;
+    let consumers = create_amqp_consumer(&conn, "consumer", config.queue.workers).await?;
+
+    spawn_workers(config.clone(), consumers).await?;
+
+    Ok(producer)
+}
+
+#[cfg(test)]
+pub fn start_test_workers() -> (Producer<MxJob>, Consumer<MxJob>) {
+    use futures::channel::mpsc::channel;
+    use monaxia_queue::queue::memory::{create_memory_consumer, create_memory_producer};
+
+    let (sender, receiver) = channel(8);
+    let producer = create_memory_producer(sender.clone());
+    let consumer = create_memory_consumer(sender, receiver);
+    (producer, consumer)
+}
+
+async fn spawn_workers(config: Arc<Config>, consumers: Vec<Consumer<MxJob>>) -> Result<()> {
     let container = construct_container_db(&config).await?;
     let http_client = create_http_client(&config)?;
 
     info!("spawning {} workers", consumers.len());
     for (i, consumer) in consumers.into_iter().enumerate() {
-        spawn(root::worker(WorkerState {
+        spawn(worker(WorkerState {
             name: format!("worker-{i}"),
             consumer,
             job_state: JobState {
@@ -52,21 +73,44 @@ pub async fn spawn_workers(config: Arc<Config>, consumers: Vec<Consumer<MxJob>>)
     Ok(())
 }
 
-pub async fn create_queues(config: &Config) -> Result<(Producer<MxJob>, Vec<Consumer<MxJob>>)> {
-    let conn = LapinConnection::connect(&config.queue.url, ConnectionProperties::default()).await?;
-    let producer = create_amqp_producer(&conn, "producer").await?;
-    let consumers = create_amqp_consumer(&conn, "consumer", config.queue.workers).await?;
+#[instrument(skip(state), fields(worker = state.name))]
+async fn worker(state: WorkerState) -> Result<()> {
+    // TODO: just loop
+    while let Some((job, delivery)) = state.consumer.fetch().await? {
+        let job_state = state.job_state.clone();
+        let job_payload = job.job().clone();
+        let job_tag = job.tag().to_string();
 
-    Ok((producer, consumers))
+        match do_job(job_state, job_payload, job_tag).await {
+            Ok(()) => {
+                state.consumer.mark_success(delivery).await?;
+            }
+            Err(e) => {
+                error!("job error: {e}");
+                state.consumer.mark_failure(delivery).await?;
+                if let Some((data, delay)) = job.next() {
+                    state.consumer.enqueue(data, Some(delay)).await?;
+                }
+            }
+        }
+    }
+
+    Ok(())
 }
 
-#[cfg(test)]
-pub fn create_test_queues() -> (Producer<MxJob>, Consumer<MxJob>) {
-    use futures::channel::mpsc::channel;
-    use monaxia_queue::queue::memory::{create_memory_consumer, create_memory_producer};
+#[instrument(
+    skip(state, payload, _tag),
+    fields(job = to_variant_name(&payload).expect("invalid job"))
+)]
+async fn do_job(state: JobState, payload: Job, _tag: String) -> Result<()> {
+    match payload {
+        Job::Hello => {
+            info!("hello monaxia!");
+        }
+        Job::ActivityPreprocess(json_text, validation) => {
+            ap::validate_request(state, json_text, validation).await?;
+        }
+    }
 
-    let (sender, receiver) = channel(8);
-    let producer = create_memory_producer(sender.clone());
-    let consumer = create_memory_consumer(sender, receiver);
-    (producer, consumer)
+    Ok(())
 }
